requirements.txt
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1
scipy>=1.11.0
